Remove unused BFci, GFci, and GraspFci2 classes
These classes have been deleted as they are no longer used or necessary. Their removal simplifies the codebase and eliminates redundancy, ensuring cleaner and more maintainable code.
