+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/GraspFci.java
@@ -23,16 +23,10 @@ package edu.cmu.tetrad.search;
 import edu.cmu.tetrad.data.Knowledge;
 import edu.cmu.tetrad.graph.*;
 import edu.cmu.tetrad.search.score.Score;
-import edu.cmu.tetrad.search.test.MsepTest;
-import edu.cmu.tetrad.search.utils.*;
-import edu.cmu.tetrad.search.work_in_progress.MagSemBicScore;
 import edu.cmu.tetrad.util.TetradLogger;
+import org.jetbrains.annotations.NotNull;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-
-import static edu.cmu.tetrad.graph.GraphUtils.gfciExtraEdgeRemovalStepUnionOfAdj;
 
 /**
  * Uses GRaSP in place of FGES for the initial step in the GFCI algorithm. This tends to produce a accurate PAG than
@@ -54,10 +48,8 @@ import static edu.cmu.tetrad.graph.GraphUtils.gfciExtraEdgeRemovalStepUnionOfAdj
  * @author josephramsey
  * @author bryanandrews
  * @version $Id: $Id
+ * @see StarFci
  * @see Grasp
- * @see GFci
- * @see FciOrient
- * @see Knowledge
  */
 public final class GraspFci implements IGraphSearch {
 
@@ -148,7 +140,7 @@ public final class GraspFci implements IGraphSearch {
      * Constructs a new GraspFci object.
      *
      * @param test  The independence test.
-     * @param score a {@link edu.cmu.tetrad.search.score.Score} object
+     * @param score a {@link Score} object
      */
     public GraspFci(IndependenceTest test, Score score) {
         if (score == null) {
@@ -178,99 +170,56 @@ public final class GraspFci implements IGraphSearch {
 
         Graph cpdag;
 
-        if (!startFromCompleteGraph) {
-            if (verbose) {
-                TetradLogger.getInstance().log("Starting GRaSP.");
-            }
-
-            // Run GRaSP to get a CPDAG (like GFCI with FGES)...
-            Grasp alg = new Grasp(independenceTest, score);
-            alg.setSeed(seed);
-            alg.setOrdered(ordered);
-            alg.setUseScore(useScore);
-            alg.setUseRaskuttiUhler(useRaskuttiUhler);
-            alg.setUseDataOrder(useDataOrder);
-            alg.setDepth(depth);
-            alg.setUncoveredDepth(uncoveredDepth);
-            alg.setNonSingularDepth(nonSingularDepth);
-            alg.setNumStarts(numStarts);
-            alg.setVerbose(false);
-            alg.setKnowledge(knowledge);
-
-            List<Node> variables = this.score.getVariables();
-            assert variables != null;
-
-            alg.bestOrder(variables);
-            cpdag = alg.getGraph(true);
-
-            if (verbose) {
-                TetradLogger.getInstance().log("Finished GRaSP.");
-            }
-
-            TetradLogger.getInstance().log("===Ending GRaSP-FCI===");
-        } else {
+        if (startFromCompleteGraph) {
             TetradLogger.getInstance().log("===Starting with complete graph=== ");
-
             cpdag = new EdgeListGraph(independenceTest.getVariables());
             cpdag = GraphUtils.completeGraph(cpdag);
-        }
-
-        Graph pag = new EdgeListGraph(cpdag);
-        pag.reorientAllWith(Endpoint.CIRCLE);
-
-        if (score instanceof MagSemBicScore) {
-            ((MagSemBicScore) score).setMag(GraphTransforms.zhangMagFromPag(pag));
-        }
-
-        SepsetProducer sepsets;
-
-        if (independenceTest instanceof MsepTest) {
-            sepsets = new DagSepsets(((MsepTest) independenceTest).getGraph());
-        } else if (sepsetFinderMethod == 1) {
-            sepsets = new SepsetsGreedy(pag, this.independenceTest, this.depth);
-        } else if (sepsetFinderMethod == 2) {
-            sepsets = new SepsetsMinP(pag, this.independenceTest, this.depth);
-        } else if (sepsetFinderMethod == 3) {
-            sepsets = new SepsetsMaxP(pag, this.independenceTest, this.depth);
         } else {
-            throw new IllegalArgumentException("Invalid sepset finder method: " + sepsetFinderMethod);
+            cpdag = getCpdag();
         }
 
-        Set<Triple> unshieldedColliders = new HashSet<>();
+        StarFci starFci = new StarFci(cpdag, independenceTest);
+        starFci.setKnowledge(knowledge);
+        starFci.setDepth(depth);
+        starFci.setSepsetFinderMethod(sepsetFinderMethod);
+        starFci.setVerbose(verbose);
+        starFci.setMaxDiscriminatingPathLength(maxDiscriminatingPathLength);
+        starFci.setGuaranteePag(guaranteePag);
+        starFci.setCompleteRuleSetUsed(completeRuleSetUsed);
 
-        gfciExtraEdgeRemovalStepUnionOfAdj(pag, cpdag, nodes, independenceTest, depth, null, verbose);
-        GraphUtils.gfciR0(pag, cpdag, sepsets, knowledge, verbose, unshieldedColliders);
+        return starFci.search();
+    }
 
-        pag.reorientAllWith(Endpoint.CIRCLE);
-        GraphUtils.gfciR0(pag, cpdag, sepsets, knowledge, verbose, unshieldedColliders);
+    private @NotNull Graph getCpdag() throws InterruptedException {
+        Graph cpdag;
+        if (verbose) {
+            TetradLogger.getInstance().log("Starting GRaSP.");
+        }
+
+        // Run GRaSP to get a CPDAG (like GFCI with FGES)...
+        Grasp alg = new Grasp(independenceTest, score);
+        alg.setSeed(seed);
+        alg.setOrdered(ordered);
+        alg.setUseScore(useScore);
+        alg.setUseRaskuttiUhler(useRaskuttiUhler);
+        alg.setUseDataOrder(useDataOrder);
+        alg.setDepth(depth);
+        alg.setUncoveredDepth(uncoveredDepth);
+        alg.setNonSingularDepth(nonSingularDepth);
+        alg.setNumStarts(numStarts);
+        alg.setVerbose(false);
+        alg.setKnowledge(knowledge);
+
+        List<Node> variables = this.score.getVariables();
+        assert variables != null;
+
+        alg.bestOrder(variables);
+        cpdag = alg.getGraph(true);
 
         if (verbose) {
-            TetradLogger.getInstance().log("Starting final FCI orientation.");
+            TetradLogger.getInstance().log("Finished GRaSP.");
         }
-
-        R0R4StrategyTestBased strategy = (R0R4StrategyTestBased) R0R4StrategyTestBased.specialConfiguration(independenceTest, knowledge, verbose);
-        strategy.setDepth(-1);
-        strategy.setMaxLength(-1);
-        FciOrient fciOrient = new FciOrient(strategy);
-        fciOrient.setCompleteRuleSetUsed(completeRuleSetUsed);
-        fciOrient.setMaxDiscriminatingPathLength(maxDiscriminatingPathLength);
-        fciOrient.setVerbose(verbose);
-
-        fciOrient.finalOrientation(pag);
-
-        if (verbose) {
-            TetradLogger.getInstance().log("Finished implied orientation.");
-        }
-
-        if (guaranteePag) {
-            pag = GraphUtils.guaranteePag(pag, fciOrient, knowledge, unshieldedColliders, unshieldedColliders, verbose, new HashSet<>());
-        }
-
-        if (verbose) {
-            TetradLogger.getInstance().log("GRaSP-FCI finished.");
-        }
-
-        return pag;
+        return cpdag;
     }
 
     /**
