--- tetrad-lib/src/main/java/edu/cmu/tetrad/search/GraspFci2.java
@@ -1,384 +1,0 @@
-/// ////////////////////////////////////////////////////////////////////////////
-// For information as to what this class does, see the Javadoc, below.       //i
-// Copyright (C) 1998, 1999, 2000, 2001, 2002, 2003, 2004, 2005, 2006,       //
-// 2007, 2008, 2009, 2010, 2014, 2015, 2022 by Peter Spirtes, Richard        //
-// Scheines, Joseph Ramsey, and Clark Glymour.                               //
-//                                                                           //
-// This program is free software; you can redistribute it and/or modify      //
-// it under the terms of the GNU General Public License as published by      //
-// the Free Software Foundation; either version 2 of the License, or         //
-// (at your option) any later version.                                       //
-//                                                                           //
-// This program is distributed in the hope that it will be useful,           //
-// but WITHOUT ANY WARRANTY; without even the implied warranty of            //
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the             //
-// GNU General Public License for more details.                              //
-//                                                                           //
-// You should have received a copy of the GNU General Public License         //
-// along with this program; if not, write to the Free Software               //
-// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
-/// ////////////////////////////////////////////////////////////////////////////
-package edu.cmu.tetrad.search;
-
-import edu.cmu.tetrad.data.Knowledge;
-import edu.cmu.tetrad.graph.*;
-import edu.cmu.tetrad.search.score.Score;
-import edu.cmu.tetrad.search.test.MsepTest;
-import edu.cmu.tetrad.search.utils.*;
-import edu.cmu.tetrad.search.work_in_progress.MagSemBicScore;
-import edu.cmu.tetrad.util.TetradLogger;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-import static edu.cmu.tetrad.graph.GraphUtils.gfciExtraEdgeRemovalStepUnionOfAdj;
-
-/**
- * Uses GRaSP in place of FGES for the initial step in the GFCI algorithm. This tends to produce a accurate PAG than
- * GFCI as a result, for the latent variables case. This is a simple substitution; the reference for GFCI is here: J.M.
- * Ogarrio and P. Spirtes and J. Ramsey, "A Hybrid Causal Search Algorithm for Latent Variable Models," JMLR 2016. Here,
- * BOSS has been substituted for FGES.
- * <p>
- * For the first step, the GRaSP algorithm is used, with the same modifications as in the GFCI algorithm.
- * <p>
- * For the second step, the FCI final orientation algorithm is used, with the same modifications as in the GFCI
- * algorithm.
- * <p>
- * For GRaSP only a score is needed, but there are steps in GFCI that require a test, so for this method, both a test
- * and a score need to be given.
- * <p>
- * This class is configured to respect knowledge of forbidden and required edges, including knowledge of temporal
- * tiers.
- *
- * @author josephramsey
- * @author bryanandrews
- * @version $Id: $Id
- * @see Grasp
- * @see GFci
- * @see FciOrient
- * @see Knowledge
- */
-public final class GraspFci2 implements IGraphSearch {
-
-    /**
-     * The conditional independence test.
-     */
-    private final IndependenceTest independenceTest;
-    /**
-     * The score.
-     */
-    private final Score score;
-    /**
-     * The background knowledge.
-     */
-    private Knowledge knowledge = new Knowledge();
-    /**
-     * Flag for the complete rule set, true if one should use the complete rule set, false otherwise.
-     */
-    private boolean completeRuleSetUsed = true;
-    /**
-     * The maximum length for any discriminating path. -1 if unlimited; otherwise, a positive integer.
-     */
-    private int maxDiscriminatingPathLength = -1;
-    /**
-     * The number of starts for GRaSP.
-     */
-    private int numStarts = 1;
-    /**
-     * Whether to use Raskutti and Uhler's modification of GRaSP.
-     */
-    private boolean useRaskuttiUhler = false;
-    /**
-     * Whether to use data order.
-     */
-    private boolean useDataOrder = true;
-    /**
-     * Whether to use score.
-     */
-    private boolean useScore = true;
-    /**
-     * Whether to use the ordered version of GRaSP.
-     */
-    private boolean ordered = false;
-    /**
-     * The depth for singular variables.
-     */
-    private int uncoveredDepth = 1;
-    /**
-     * The depth for non-singular variables.
-     */
-    private int nonSingularDepth = 1;
-    /**
-     * The depth for sepsets.
-     */
-    private int depth = -1;
-    /**
-     * The seed used for random number generation. If the seed is not set explicitly, it will be initialized with a
-     * value of -1. The seed is used for producing the same sequence of random numbers every time the program runs.
-     *
-     * @see GraspFci2#setSeed(long)
-     */
-    private long seed = -1;
-    /**
-     * True iff verbose output should be printed.
-     */
-    private boolean verbose = false;
-    /**
-     * The flag for whether to guarantee the output is a legal PAG.
-     */
-    private boolean guaranteePag = false;
-    /**
-     * The method to use for finding sepsets, 1 = greedy, 2 = min-p., 3 = max-p, default min-p.
-     */
-    private int sepsetFinderMethod = 2;
-    /**
-     * Determines whether the GRaSP-FCI algorithm should start with a complete graph.
-     * <p>
-     * If set to true, the algorithm begins with a fully connected graph, where every variable is connected to every
-     * other variable. If set to false, the algorithm starts with an empty graph or another specified initial
-     * structure.
-     * <p>
-     * This option can impact the performance and results of the search, as the starting graph influences the edges to
-     * be considered or removed during the causal discovery process.
-     */
-    private boolean startFromCompleteGraph = false;
-
-    /**
-     * Constructs a new GraspFci object.
-     *
-     * @param test  The independence test.
-     * @param score a {@link Score} object
-     */
-    public GraspFci2(IndependenceTest test, Score score) {
-        if (score == null) {
-            throw new NullPointerException();
-        }
-
-        this.score = score;
-        this.independenceTest = test;
-    }
-
-    /**
-     * Run the search and return s a PAG.
-     *
-     * @return The PAG.
-     * @throws InterruptedException if any
-     */
-    public Graph search() throws InterruptedException {
-        List<Node> nodes = this.independenceTest.getVariables();
-
-        if (nodes == null) {
-            throw new NullPointerException("Nodes from test were null.");
-        }
-
-        if (verbose) {
-            TetradLogger.getInstance().log("===Starting GRaSP-FCI===");
-        }
-
-        Graph cpdag;
-
-        if (startFromCompleteGraph) {
-            TetradLogger.getInstance().log("===Starting with complete graph=== ");
-            cpdag = new EdgeListGraph(independenceTest.getVariables());
-            cpdag = GraphUtils.completeGraph(cpdag);
-        } else {
-            cpdag = getCpdag();
-        }
-
-        StarFci starFci = new StarFci(cpdag, independenceTest);
-        starFci.setKnowledge(knowledge);
-        starFci.setDepth(depth);
-        starFci.setSepsetFinderMethod(sepsetFinderMethod);
-        starFci.setVerbose(verbose);
-        starFci.setMaxDiscriminatingPathLength(maxDiscriminatingPathLength);
-        starFci.setGuaranteePag(guaranteePag);
-        starFci.setCompleteRuleSetUsed(completeRuleSetUsed);
-
-        return starFci.search();
-    }
-
-    private @NotNull Graph getCpdag() throws InterruptedException {
-        Graph cpdag;
-        if (verbose) {
-            TetradLogger.getInstance().log("Starting GRaSP.");
-        }
-
-        // Run GRaSP to get a CPDAG (like GFCI with FGES)...
-        Grasp alg = new Grasp(independenceTest, score);
-        alg.setSeed(seed);
-        alg.setOrdered(ordered);
-        alg.setUseScore(useScore);
-        alg.setUseRaskuttiUhler(useRaskuttiUhler);
-        alg.setUseDataOrder(useDataOrder);
-        alg.setDepth(depth);
-        alg.setUncoveredDepth(uncoveredDepth);
-        alg.setNonSingularDepth(nonSingularDepth);
-        alg.setNumStarts(numStarts);
-        alg.setVerbose(false);
-        alg.setKnowledge(knowledge);
-
-        List<Node> variables = this.score.getVariables();
-        assert variables != null;
-
-        alg.bestOrder(variables);
-        cpdag = alg.getGraph(true);
-
-        if (verbose) {
-            TetradLogger.getInstance().log("Finished GRaSP.");
-        }
-        return cpdag;
-    }
-
-    /**
-     * Sets the knowledge used in search.
-     *
-     * @param knowledge This knowledge.
-     */
-    public void setKnowledge(Knowledge knowledge) {
-        this.knowledge = new Knowledge(knowledge);
-    }
-
-    /**
-     * Sets whether Zhang's complete rules set is used.
-     *
-     * @param completeRuleSetUsed set to true if Zhang's complete rule set should be used, false if only R1-R4 (the rule
-     *                            set of the original FCI) should be used. False by default.
-     */
-    public void setCompleteRuleSetUsed(boolean completeRuleSetUsed) {
-        this.completeRuleSetUsed = completeRuleSetUsed;
-    }
-
-    /**
-     * Sets the maximum length of any discriminating path.
-     *
-     * @param maxDiscriminatingPathLength the maximum length of any discriminating path, or -1 if unlimited.
-     */
-    public void setMaxDiscriminatingPathLength(int maxDiscriminatingPathLength) {
-        if (maxDiscriminatingPathLength < -1) {
-            throw new IllegalArgumentException("Max path length must be -1 (unlimited) or >= 0: " + maxDiscriminatingPathLength);
-        }
-
-        this.maxDiscriminatingPathLength = maxDiscriminatingPathLength;
-    }
-
-    /**
-     * Sets whether verbose output should be printed.
-     *
-     * @param verbose True, if so.
-     */
-    public void setVerbose(boolean verbose) {
-        this.verbose = verbose;
-    }
-
-    /**
-     * Sets the number of starts for GRaSP.
-     *
-     * @param numStarts The number of starts.
-     */
-    public void setNumStarts(int numStarts) {
-        this.numStarts = numStarts;
-    }
-
-    /**
-     * Sets whether to use Raskutti and Uhler's modification of GRaSP.
-     *
-     * @param useRaskuttiUhler True, if so.
-     */
-    public void setUseRaskuttiUhler(boolean useRaskuttiUhler) {
-        this.useRaskuttiUhler = useRaskuttiUhler;
-    }
-
-    /**
-     * Sets whether to use data order for GRaSP (as opposed to random order) for the first step of GRaSP
-     *
-     * @param useDataOrder True, if so.
-     */
-    public void setUseDataOrder(boolean useDataOrder) {
-        this.useDataOrder = useDataOrder;
-    }
-
-    /**
-     * Sets whether to use score for GRaSP (as opposed to independence test) for GRaSP.
-     *
-     * @param useScore True, if so.
-     */
-    public void setUseScore(boolean useScore) {
-        this.useScore = useScore;
-    }
-
-    /**
-     * Sets depth for singular tucks.
-     *
-     * @param uncoveredDepth The depth for singular tucks.
-     */
-    public void setSingularDepth(int uncoveredDepth) {
-        if (uncoveredDepth < -1) throw new IllegalArgumentException("Uncovered depth should be >= -1.");
-        this.uncoveredDepth = uncoveredDepth;
-    }
-
-    /**
-     * Sets depth for non-singular tucks.
-     *
-     * @param nonSingularDepth The depth for non-singular tucks.
-     */
-    public void setNonSingularDepth(int nonSingularDepth) {
-        if (nonSingularDepth < -1) throw new IllegalArgumentException("Non-singular depth should be >= -1.");
-        this.nonSingularDepth = nonSingularDepth;
-    }
-
-    /**
-     * Sets whether to use the ordered version of GRaSP.
-     *
-     * @param ordered True, if so.
-     */
-    public void setOrdered(boolean ordered) {
-        this.ordered = ordered;
-    }
-
-    /**
-     * <p>Setter for the field <code>seed</code>.</p>
-     *
-     * @param seed a long
-     */
-    public void setSeed(long seed) {
-        this.seed = seed;
-    }
-
-    /**
-     * Sets the depth for the search algorithm.
-     *
-     * @param depth The depth value to set for the search algorithm.
-     */
-    public void setDepth(int depth) {
-        this.depth = depth;
-    }
-
-    /**
-     * Sets the flag for whether to guarantee the output is a legal PAG.
-     *
-     * @param guaranteePag True, if so.
-     */
-    public void setGuaranteePag(boolean guaranteePag) {
-        this.guaranteePag = guaranteePag;
-    }
-
-    /**
-     * Sets the method for finding sepsets in the GraspFci class.
-     *
-     * @param sepsetFinderMethod the method for finding sepsets
-     */
-    public void setSepsetFinderMethod(int sepsetFinderMethod) {
-        this.sepsetFinderMethod = sepsetFinderMethod;
-    }
-
-    /**
-     * If true, the score search is skipped and the search starts with a complete graph. Default is false.
-     *
-     * @param startFromCompleteGraph true if the search should start with a complete graph, false otherwise.
-     */
-    public void setStartFromCompleteGraph(boolean startFromCompleteGraph) {
-        this.startFromCompleteGraph = startFromCompleteGraph;
-    }
-}
