+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/SpFci.java
@@ -23,18 +23,10 @@ package edu.cmu.tetrad.search;
 import edu.cmu.tetrad.data.Knowledge;
 import edu.cmu.tetrad.graph.*;
 import edu.cmu.tetrad.search.score.Score;
-import edu.cmu.tetrad.search.test.MsepTest;
-import edu.cmu.tetrad.search.utils.*;
-import edu.cmu.tetrad.search.work_in_progress.MagSemBicScore;
 import edu.cmu.tetrad.util.TetradLogger;
 
 import java.io.PrintStream;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-
-import static edu.cmu.tetrad.graph.GraphUtils.gfciExtraEdgeRemovalStep;
-import static edu.cmu.tetrad.graph.GraphUtils.gfciExtraEdgeRemovalStepUnionOfAdj;
 
 /**
  * Uses SP in place of FGES for the initial step in the GFCI algorithm. This tends to produce an accurate PAG than GFCI
@@ -60,11 +52,8 @@ import static edu.cmu.tetrad.graph.GraphUtils.gfciExtraEdgeRemovalStepUnionOfAdj
  * @author josephramsey
  * @author bryan andrews
  * @version $Id: $Id
- * @see Grasp
- * @see GFci
+ * @see StarFci
  * @see Sp
- * @see Knowledge
- * @see FciOrient
  */
 public final class SpFci implements IGraphSearch {
 
@@ -138,88 +127,53 @@ public final class SpFci implements IGraphSearch {
      * @return This PAG.
      */
     public Graph search() throws InterruptedException {
-        List<Node> nodes = getIndependenceTest().getVariables();
+        List<Node> nodes = this.independenceTest.getVariables();
+
+        if (nodes == null) {
+            throw new NullPointerException("Nodes from test were null.");
+        }
 
         if (verbose) {
-            TetradLogger.getInstance().log("===Starting SP-FCI===");
+            TetradLogger.getInstance().log("===Starting GRaSP-FCI===");
         }
 
         Graph cpdag;
 
-        if (!startFromCompleteGraph) {
-            if (verbose) {
-                TetradLogger.getInstance().log("Starting SP.");
-            }
-
-            Sp subAlg = new Sp(this.score);
-            PermutationSearch alg = new PermutationSearch(subAlg);
-            alg.setKnowledge(this.knowledge);
-            cpdag = alg.search();
-
-            if (verbose) {
-                TetradLogger.getInstance().log("Finished SP.");
-            }
-        } else {
+        if (startFromCompleteGraph) {
             TetradLogger.getInstance().log("===Starting with complete graph=== ");
             cpdag = new EdgeListGraph(independenceTest.getVariables());
             cpdag = GraphUtils.completeGraph(cpdag);
-        }
-
-        Graph pag = new EdgeListGraph(cpdag);
-        pag.reorientAllWith(Endpoint.CIRCLE);
-
-        if (score instanceof MagSemBicScore) {
-            ((MagSemBicScore) score).setMag(pag);
-        }
-
-        SepsetProducer sepsets;
-
-        if (independenceTest instanceof MsepTest) {
-            sepsets = new DagSepsets(((MsepTest) independenceTest).getGraph());
-        } else if (sepsetFinderMethod == 1) {
-            sepsets = new SepsetsGreedy(pag, this.independenceTest, this.depth);
-        } else if (sepsetFinderMethod == 2) {
-            sepsets = new SepsetsMinP(pag, this.independenceTest, this.depth);
-        } else if (sepsetFinderMethod == 3) {
-            sepsets = new SepsetsMaxP(pag, this.independenceTest, this.depth);
         } else {
-            throw new IllegalArgumentException("Invalid sepset finder method: " + sepsetFinderMethod);
+            cpdag = getCpdag();
         }
 
-        Set<Triple> unshieldedColliders = new HashSet<>();
+        StarFci starFci = new StarFci(cpdag, independenceTest);
+        starFci.setKnowledge(knowledge);
+        starFci.setDepth(depth);
+        starFci.setSepsetFinderMethod(sepsetFinderMethod);
+        starFci.setVerbose(verbose);
+        starFci.setMaxDiscriminatingPathLength(maxDiscriminatingPathLength);
+        starFci.setGuaranteePag(guaranteePag);
+        starFci.setCompleteRuleSetUsed(completeRuleSetUsed);
 
-        gfciExtraEdgeRemovalStepUnionOfAdj(pag, cpdag, nodes, independenceTest, depth, null, verbose);
-        GraphUtils.gfciR0(pag, cpdag, sepsets, knowledge, verbose, unshieldedColliders);
+        return starFci.search();
+    }
+
+    private Graph getCpdag() throws InterruptedException {
+        Graph cpdag;
+        if (verbose) {
+            TetradLogger.getInstance().log("Starting SP.");
+        }
+
+        Sp subAlg = new Sp(this.score);
+        PermutationSearch alg = new PermutationSearch(subAlg);
+        alg.setKnowledge(this.knowledge);
+        cpdag = alg.search();
 
         if (verbose) {
-            TetradLogger.getInstance().log("Starting final FCI orientation.");
+            TetradLogger.getInstance().log("Finished SP.");
         }
-
-        R0R4StrategyTestBased strategy = (R0R4StrategyTestBased) R0R4StrategyTestBased.specialConfiguration(independenceTest,
-                knowledge, verbose);
-        strategy.setDepth(-1);
-        strategy.setMaxLength(-1);
-        FciOrient fciOrient = new FciOrient(strategy);
-        fciOrient.setCompleteRuleSetUsed(completeRuleSetUsed);
-        fciOrient.setMaxDiscriminatingPathLength(maxDiscriminatingPathLength);
-        fciOrient.setVerbose(verbose);
-
-        fciOrient.finalOrientation(pag);
-
-        if (verbose) {
-            TetradLogger.getInstance().log("Finished implied orientation.");
-        }
-
-        if (guaranteePag) {
-            pag = GraphUtils.guaranteePag(pag, fciOrient, knowledge, unshieldedColliders, unshieldedColliders, verbose,
-                    new HashSet<>());
-        }
-
-        if (verbose) {
-            TetradLogger.getInstance().log("SP-FCI finished.");
-        }
-
-        return pag;
+        return cpdag;
     }
 
     /**
