+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/utils/FciOrient.java
@@ -1,4 +1,4 @@
-///////////////////////////////////////////////////////////////////////////////
+/// ////////////////////////////////////////////////////////////////////////////
 // For information as to what this class does, see the Javadoc, below.       //
 // Copyright (C) 1998, 1999, 2000, 2001, 2002, 2003, 2004, 2005, 2006,       //
 // 2007, 2008, 2009, 2010, 2014, 2015, 2022 by Peter Spirtes, Richard        //
@@ -23,10 +23,7 @@ package edu.cmu.tetrad.search.utils;
 import edu.cmu.tetrad.data.Knowledge;
 import edu.cmu.tetrad.data.KnowledgeEdge;
 import edu.cmu.tetrad.graph.*;
-import edu.cmu.tetrad.search.Fci;
 import edu.cmu.tetrad.search.FciOrientDijkstra;
-import edu.cmu.tetrad.search.GFci;
-import edu.cmu.tetrad.search.Rfci;
 import edu.cmu.tetrad.util.ChoiceGenerator;
 import edu.cmu.tetrad.util.TetradLogger;
 import org.apache.commons.lang3.tuple.Pair;
@@ -67,9 +64,6 @@ import java.util.concurrent.TimeUnit;
  * @author josephramsey 2024-8-21
  * @author Choh-Man Teng
  * @version $Id: $Id
- * @see Fci
- * @see GFci
- * @see Rfci
  */
 public class FciOrient {
