+++ tetrad-lib/src/main/java/edu/cmu/tetrad/study/performance/PerformanceTests.java
@@ -861,7 +861,7 @@ public class PerformanceTests {
 
         SemBicScore score = new SemBicScore(cov);
         score.setPenaltyDiscount(penaltyDiscount);
-        GFci fci = new GFci(independenceTest, score);
+        Gfci fci = new Gfci(independenceTest, score);
 
         fci.setVerbose(false);
         fci.setMaxDiscriminatingPathLength(maxDiscriminatingPathLength);
@@ -1526,7 +1526,7 @@ public class PerformanceTests {
             long ta1 = MillisecondTimes.timeMillis();
 
 //            FCI fci = new FCI(independenceTest);
-            GFci fci = new GFci(independenceTest, score);
+            Gfci fci = new Gfci(independenceTest, score);
             fci.setMaxDegree(depth);
             fci.setMaxDiscriminatingPathLength(maxPathLength);
 //            fci.setPossibleNsepSearchDone(doPossibleDsep);
