+++ tetrad-lib/src/main/java/edu/cmu/tetrad/study/performance/PerformanceTestsDan.java
@@ -1,4 +1,4 @@
-///////////////////////////////////////////////////////////////////////////////
+/// ////////////////////////////////////////////////////////////////////////////
 // For information as to what this class does, see the Javadoc, below.       //
 // Copyright (C) 1998, 1999, 2000, 2001, 2002, 2003, 2004, 2005, 2006,       //
 // 2007, 2008, 2009, 2010, 2014, 2015, 2022 by Peter Spirtes, Richard        //
@@ -26,7 +26,7 @@ import edu.cmu.tetrad.data.DataSet;
 import edu.cmu.tetrad.data.DataTransforms;
 import edu.cmu.tetrad.data.ICovarianceMatrix;
 import edu.cmu.tetrad.graph.*;
-import edu.cmu.tetrad.search.GFci;
+import edu.cmu.tetrad.search.Gfci;
 import edu.cmu.tetrad.search.Pc;
 import edu.cmu.tetrad.search.score.SemBicScore;
 import edu.cmu.tetrad.search.test.IndTestFisherZ;
@@ -197,7 +197,7 @@ public class PerformanceTestsDan {
 
             out6.println("GFCI.PAG_of_the_true_DAG");
 
-            GFci gFci = new GFci(independenceTestGFci, scoreGfci);
+            Gfci gFci = new Gfci(independenceTestGFci, scoreGfci);
             gFci.setVerbose(false);
             gFci.setMaxDegree(depth);
             gFci.setMaxDiscriminatingPathLength(maxDiscriminatingPathLength);
