+++ tetrad-lib/src/test/java/edu/cmu/tetrad/test/TestPc.java
@@ -340,7 +340,7 @@ public class TestPc {
                     search = new Fci(test);
                     break;
                 case 4:
-                    search = new GFci(test, score);
+                    search = new Gfci(test, score);
                     break;
                 case 5:
                     search = new Rfci(test);
@@ -707,7 +707,7 @@ public class TestPc {
                     }
                     break;
                 case 4:
-                    search = new GFci(test, score);
+                    search = new Gfci(test, score);
                     try {
                         out = search.search();
                     } catch (InterruptedException e) {
